--- models/db.py
@@ -1,5 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
-
-instance = "mysql+pymysql://lucas:lucas@localhost:3306/puhl_gasperin_health"
